arc_zkp::sigma

Trait SigmaProtocol

pub trait SigmaProtocol {
    type Statement;
    type Witness;
    type Commitment;
    type Response;

    // Required methods
    fn commit(
        &self,
        statement: &Self::Statement,
        witness: &Self::Witness,
    ) -> Result<(Self::Commitment, Vec<u8>)>;
    fn respond(
        &self,
        witness: &Self::Witness,
        commitment_state: Vec<u8>,
        challenge: &[u8; 32],
    ) -> Result<Self::Response>;
    fn verify(
        &self,
        statement: &Self::Statement,
        commitment: &Self::Commitment,
        challenge: &[u8; 32],
        response: &Self::Response,
    ) -> Result<bool>;
    fn serialize_commitment(&self, commitment: &Self::Commitment) -> Vec<u8> ;
    fn deserialize_commitment(&self, bytes: &[u8]) -> Result<Self::Commitment>;
    fn serialize_response(&self, response: &Self::Response) -> Vec<u8> ;
    fn deserialize_response(&self, bytes: &[u8]) -> Result<Self::Response>;
    fn serialize_statement(&self, statement: &Self::Statement) -> Vec<u8> ;
}

Trait for implementing sigma protocols

Required Associated Types

type Statement

Statement type (what we’re proving)

type Witness

Witness type (the secret)

type Commitment

Commitment type

type Response

Response type

Required Methods

fn commit( &self, statement: &Self::Statement, witness: &Self::Witness, ) -> Result<(Self::Commitment, Vec<u8>)>

Generate commitment (step 1)

Errors

Returns an error if commitment generation fails.

fn respond( &self, witness: &Self::Witness, commitment_state: Vec<u8>, challenge: &[u8; 32], ) -> Result<Self::Response>

Generate response given challenge (step 3)

Errors

Returns an error if response computation fails.

fn verify( &self, statement: &Self::Statement, commitment: &Self::Commitment, challenge: &[u8; 32], response: &Self::Response, ) -> Result<bool>

Verify the proof

Errors

Returns an error if proof verification fails due to invalid data.

fn serialize_commitment(&self, commitment: &Self::Commitment) -> Vec<u8>

Serialize commitment for Fiat-Shamir

fn deserialize_commitment(&self, bytes: &[u8]) -> Result<Self::Commitment>

Deserialize commitment

Errors

Returns an error if the bytes do not represent a valid commitment.

fn serialize_response(&self, response: &Self::Response) -> Vec<u8>

Serialize response

fn deserialize_response(&self, bytes: &[u8]) -> Result<Self::Response>

Deserialize response

Errors

Returns an error if the bytes do not represent a valid response.

fn serialize_statement(&self, statement: &Self::Statement) -> Vec<u8>

Serialize statement for challenge computation